LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/prefetch-single.https_from_protocol=https&to_protocol=http-expected.txt

CONSOLE MESSAGE: Prefetch request denied: not same origin as document

PASS test single http url prefetch from a https url
